MOD-13409 Share RDB corruption test helpers
